requirements.txt
tqdm
python-docx
openai
tiktoken
